refactor: Move remainders of `GitRepoFunTest` to `AnalyzerFunTest`
The remaining test in `GitRepoFunTest` is sort of an integration test that
combines testing the downloader with multiple package managers from the
analyzer / plugins. As such it is better hosted in the `cli` module that
bundles these module dependencies.
