+++ .github/workflows/build-and-test.yml
@@ -131,7 +131,7 @@ jobs:
     - name: Run functional tests that do not require external tools
       uses: gradle/gradle-build-action@v2
       with:
-        arguments: --scan -Ptests.exclude=org.ossreviewtoolkit.analyzer.*,org.ossreviewtoolkit.plugins.packagemanagers.* funTest jacocoFunTestReport
+        arguments: --scan -Ptests.exclude=org.ossreviewtoolkit.cli.*,org.ossreviewtoolkit.analyzer.*,org.ossreviewtoolkit.plugins.packagemanagers.* funTest jacocoFunTestReport
     - name: Upload code coverage data
       uses: codecov/codecov-action@v3
       with:
