+++ cli/src/funTest/assets/git-repo-expected-output.yml
